fix broken picture & organize compoents
