+++ front-end/src/components/createvocab/NewCard.js
