+++ front-end/src/components/flashcard/alphabet/Alphabet.css
