+++ front-end/src/components/flashcard/alphabet/AlphabetCard.js
