+++ front-end/src/components/flashcard/vocab/VocabCard.js
