+++ front-end/src/components/home/CardStyle.css
