+++ front-end/src/components/quizes/Categories.js
@@ -26,7 +26,7 @@ const Categories = [
     title: "Clima",
     description: " Aprenda vocabulario esencial sobre el Clima y el tiempo ",
     img:
-      "https://cdn.pixabay.com/photo/2014/05/18/00/33/cloud-346710_960_720.png",
+      "https://cdn.pixabay.com/photo/2017/06/03/10/15/bulb-2368396_1280.jpg",
     imgTitle: "Weather Forecast",
   },
   {
