+++ front-end/src/components/quizes/payquiz/Quiz.css
