+++ front-end/src/components/quizes/payquiz/Quiz.js
