+++ front-end/src/components/quizes/payquiz/QuizInfo.js
