+++ front-end/src/components/quizes/payquiz/Quizparent.js
