+++ front-end/src/components/translation/EnglishRec.js
