+++ front-end/src/components/translation/Translate.js
